agentic_os/requirements.txt
## Recommended runtime: Python 3.10 (validated). Avoid 3.14 for now due ecosystem compatibility.
langgraph
langchain
langchain-mistralai
langchain-community
fastapi
uvicorn[standard]
python-dotenv
duckduckgo-search
psutil
requests
httpx
sentence-transformers
faiss-cpu
pyautogui
pillow
aiohttp
numpy
pandas
scikit-learn
ddgs
